% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw the graph associated to a name}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{name you're looking for}

\item{the_sex}{sex you're looking for}
}
\value{
a graph
}
\description{
Draw the graph associated to a name
}
\examples{
\dontrun{
draw_a_name("eric")
}
}
