% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Compare the popularity of two names}
\usage{
draw_names(names)
}
\arguments{
\item{names}{names you're interested in comparing}
}
\value{
a graph
}
\description{
Compare the popularity of two names
}
\examples{
\dontrun{
draw_names("thomas, margaux")
}
}
