% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Sava a dataset in a csv file, output full csv path}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{name of the data frame}

\item{filename}{name of the csv file the file will be saved as}

\item{row.names}{names of the rows}

\item{...}{more parameters}
}
\value{
path to csv file
}
\description{
Sava a dataset in a csv file, output full csv path
}
\examples{
\dontrun{
save_as_csv(iris, "outputiris.csv", row.names=FALSE, ...)
}

}
